+++ packages/components/form-renderer/demos/basic.vue
@@ -6,7 +6,9 @@
         <a-button @click="resetForm">重置</a-button>
       </a-space>
     </div>
-    <vz-form-renderer ref="formRenderer" @day-checkboxChange="dayCheckboxChange" :data="list"></vz-form-renderer>
+    {{ list }}
+    <vz-form-renderer @tb-name-change="onNameChange" ref="formRenderer" @day-checkboxChange="dayCheckboxChange"
+      :data="list"></vz-form-renderer>
   </div>
 </template>
 
@@ -26,6 +28,10 @@ onMounted(() => {
   console.info("formRenderer =>", formRenderer.value);
 });
 
+const onNameChange = (data: any) => {
+  console.info('onNameChange 透传事情 监听 nbnb => xxxxxxxxxxxxx', data);
+}
+
 const getFormData = () => {
   console.info("getFormData =>", formRenderer.value.formInstance.modelRef);
 };
@@ -34,14 +40,14 @@ const resetForm = () => {
   formRenderer.value.formInstance.resetFields();
 };
 
-const dayCheckboxChange = ({checked}:{checked:boolean}) => {
-  console.info('dayCheckboxChange.checked =>',checked);
+const dayCheckboxChange = ({ checked }: { checked: boolean }) => {
+  console.info('dayCheckboxChange.checked =>', checked);
   // expired
   formRenderer.value.formInstance.modelRef.expired = checked;
-  console.info('formRenderer.value.formInstance.modelRef =>',formRenderer.value.formInstance.modelRef);
+  console.info('formRenderer.value.formInstance.modelRef =>', formRenderer.value.formInstance.modelRef);
 
-  
-  
+
+
 }
 
 </script>