+++ packages/components/monaco-editor/src/index.vue
@@ -27,10 +27,15 @@ import {
 
 import * as monaco from "monaco-editor";
 
+//@ts-ignore
 import jsonWorker from "monaco-editor/esm/vs/language/json/json.worker?worker";
+//@ts-ignore
 import cssWorker from "monaco-editor/esm/vs/language/css/css.worker?worker";
+//@ts-ignore
 import htmlWorker from "monaco-editor/esm/vs/language/html/html.worker?worker";
+//@ts-ignore
 import tsWorker from "monaco-editor/esm/vs/language/typescript/ts.worker?worker";
+//@ts-ignore
 import editorWorker from "monaco-editor/esm/vs/editor/editor.worker?worker";
 
 let editor: monaco.editor.IStandaloneCodeEditor | undefined;