+++ packages/components/page/src/index.vue
@@ -31,8 +31,6 @@
   }">
     <div :class="`${prefixCls}-filter`" v-if="!initialing" ref="filterEl">
       <VzFormRenderer :data="searcher"></VzFormRenderer>
-
-      <div>searcher</div>
       <!-- aa -->
 
       <!-- <vz-form -->