+++ packages/hooks/src/useMessage.tsx
@@ -157,3 +157,6 @@ export function useMessage() {
     createWarningModal,
   };
 }
+
+
+export default useMessage;