+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @viaz/constants
 
+## 0.0.2-beta.10
+
+### Patch Changes
+
+- update import entry
+
+## 0.0.2-beta.9
+
+### Patch Changes
+
+- remove vz-page searcher
+
 ## 0.0.2-beta.8
 
 ### Patch Changes