+++ packages/enums/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @imsjs/ims-ui-enums
 
+## 0.0.2-beta.10
+
+### Patch Changes
+
+- update import entry
+- Updated dependencies
+  - @viaz/hooks@0.0.2-beta.10
+
+## 0.0.2-beta.9
+
+### Patch Changes
+
+- remove vz-page searcher
+- Updated dependencies
+  - @viaz/hooks@0.0.2-beta.9
+
 ## 0.0.2-beta.8
 
 ### Patch Changes