+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @viaz/types
 
+## 0.0.2-beta.10
+
+### Patch Changes
+
+- update import entry
+- Updated dependencies
+  - @viaz/constants@0.0.2-beta.10
+  - @viaz/enums@0.0.2-beta.10
+
+## 0.0.2-beta.9
+
+### Patch Changes
+
+- remove vz-page searcher
+- Updated dependencies
+  - @viaz/constants@0.0.2-beta.9
+  - @viaz/enums@0.0.2-beta.9
+
 ## 0.0.2-beta.8
 
 ### Patch Changes