+++ packages/viaz/CHANGELOG.md
@@ -1,5 +1,31 @@
 # viaz
 
+## 0.0.2-beta.10
+
+### Patch Changes
+
+- update import entry
+- Updated dependencies
+  - @viaz/components@0.0.2-beta.10
+  - @viaz/directives@0.0.2-beta.10
+  - @viaz/constants@0.0.2-beta.10
+  - @viaz/hooks@0.0.2-beta.10
+  - @viaz/types@0.0.2-beta.10
+  - @viaz/utils@0.0.2-beta.10
+
+## 0.0.2-beta.9
+
+### Patch Changes
+
+- remove vz-page searcher
+- Updated dependencies
+  - @viaz/components@0.0.2-beta.9
+  - @viaz/directives@0.0.2-beta.9
+  - @viaz/constants@0.0.2-beta.9
+  - @viaz/hooks@0.0.2-beta.9
+  - @viaz/types@0.0.2-beta.9
+  - @viaz/utils@0.0.2-beta.9
+
 ## 0.0.2-beta.8
 
 ### Patch Changes